binaries use minimal main
help message shows correct command name
