+++ src/xargo.rs
@@ -20,15 +20,11 @@ pub fn run(
     rustflags: Rustflags,
     home: &Home,
     meta: &VersionMeta,
-    use_rustc: bool,
+    command_name: &str,
     verbose: bool,
 ) -> Result<ExitStatus> {
     let mut cmd = Command::new("cargo");
-    if use_rustc {
-        cmd.arg("rustc");
-    } else {
-        cmd.arg("build");
-    }
+    cmd.arg(command_name);
     cmd.args(args.all());
 
     let flags = rustflags.for_xargo(home);
